requirements.txt
flask
qrcode
pillow
fpdf
pandas
openpyxl
Pillow
pylibdmtx
numpy
setuptools